Reorganise and rebuild project for FitNus to be main class
