+++ src/main/java/seedu/fitnus/CSVReader.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.fitnus;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
