+++ src/main/java/seedu/fitnus/exception/IncompleteMealException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exception;
+package seedu.fitnus.exception;
 
 public class IncompleteMealException extends Exception{
 }
