+++ src/main/java/seedu/fitnus/exception/IncompleteWaterException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exception;
+package seedu.fitnus.exception;
 
 public class IncompleteWaterException extends Exception{
 }
