+++ src/main/java/seedu/fitnus/exception/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exception;
+package seedu.fitnus.exception;
 
 public class InvalidCommandException extends Exception{
 }
