+++ src/main/java/seedu/fitnus/exception/UnregisteredDrinkException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exception;
+package seedu.fitnus.exception;
 
 public class UnregisteredDrinkException extends Exception{
 }
