+++ src/main/java/seedu/fitnus/exception/UnregisteredMealException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exception;
+package seedu.fitnus.exception;
 
 public class UnregisteredMealException extends Exception{
 }
